requirements/base.txt
Django==2.2.16
django-solo==1.1.5
django-environ==0.4.5
pytz==2020.5
sqlparse==0.4.1
Pillow==8.1.0
psycopg2-binary==2.8.5
django-2gis-maps==1.0.3
django-ckeditor==6.0.0
djangorestframework==3.12.4
drf-yasg
fcm-django==0.3.7
xmltodict==0.12.0
dicttoxml==1.7.4
factory-boy==3.2.0
haversine==2.3.0
django-filter==2.4.0
django-jet==1.0.8
django-admin-sortable2==0.7.8
huey==2.3.0
redis==3.5.3
loguru==0.5.3
